+++ R/statistic-wmean.R
@@ -1,15 +1,15 @@
 #' Weighted Mean (WMEAN)
 #'
 #' @description
-#' Calculates regulatory activities by computing the WMEAN.
+#' Calculates regulatory activities using WMEAN.
 #'
 #' @details
-#' Infers activity score for each regulator by weighting the molecular readouts
-#' of its targets by their mode of regulations and likelihoods. In addition, it
-#' runs permutations to calculate empirical p-values, providing normalized
-#' (z-score) and corrected activity (estimate * -log10(pval)) scores. This is
-#' represented in the `statistic` column which will contain three values for
-#' each call to `run_wmean()`; __wmean__, __norm_wmean__ and __corr_wmean__.
+#' WMEAN infers regulator activities by first multiplying each target feature by
+#' its associated weight which then are summed to an enrichment score
+#' `wmean`. Furthermore, permutations of random target features can
+#' be performed to obtain a null distribution that can be used to compute a
+#' z-score `norm_wmean`, or a corrected estimate `corr_wmean` by multiplying
+#' `wmean` by the minus log10 of the obtained empirical p-value.
 #'
 #' @inheritParams .decoupler_mat_format
 #' @inheritParams .decoupler_network_format
@@ -153,7 +153,7 @@ run_wmean <- function(mat,
             p_value = if_else(.data$p_value == 0, 1/times, .data$p_value),
             p_value = if_else(.data$p_value == 1, (times-1)/times, .data$p_value),
             p_value = if_else(.data$p_value >= 0.5, 1-.data$p_value, .data$p_value),
-            p_value = p_value * 2,
+            p_value = .data$p_value * 2,
             c_score = .data$value * (-log10(.data$p_value))
         ) %>%
         # Reformat results