+++ R/statistic-wsum.R
@@ -1,15 +1,15 @@
 #' Weighted Sum (WSUM)
 #'
 #' @description
-#' Calculates regulatory activities by computing the WSUM
+#' Calculates regulatory activities using WSUM.
 #'
 #' @details
-#' Infers activity score for each regulator by weighting the molecular readouts
-#' of its targets by their mode of regulations and likelihoods. In addition, it
-#' runs permutations to calculate empirical p-values, providing normalized
-#' (z-score) and corrected activity (estimate * -log10(p-value)) scores. This is
-#' represented in the `statistic` column which will contain three values for
-#' each call to `run_wsum()`; __wsum__, __norm_wsum__ and __corr_wsum__.
+#' WSUM infers regulator activities by first multiplying each target feature by
+#' its associated weight which then are summed to an enrichment score
+#' `wsum`. Furthermore, permutations of random target features can be
+#' performed to obtain a null distribution that can be used to compute a z-score
+#' `norm_wsum`, or a corrected estimate `corr_wsum` by multiplying
+#' `wsum` by the minus log10 of the obtained empirical p-value.
 #'
 #' @inheritParams .decoupler_mat_format
 #' @inheritParams .decoupler_network_format
@@ -150,7 +150,7 @@ run_wsum <- function(mat,
             p_value = if_else(.data$p_value == 0, 1/times, .data$p_value),
             p_value = if_else(.data$p_value == 1, (times-1)/times, .data$p_value),
             p_value = if_else(.data$p_value >= 0.5, 1-.data$p_value, .data$p_value),
-            p_value = p_value * 2,
+            p_value = .data$p_value * 2,
             c_score = .data$value * (-log10(.data$p_value))
         ) %>%
         # Reformat results