+++ inst/testdata/generate_expected_output.R
@@ -2,11 +2,7 @@ library(decoupleR)
 library(purrr)
 
 # Base directories definition --------------------------------------------------
-
-# Inputs
-input_dir <- system.file("testdata", "inputs", package = "decoupleR")
-
-# Outputs
+input_dir <- file.path("inst", "testdata", "inputs")
 output_dir <- file.path("inst", "testdata", "outputs")
 
 # Specific directories creation ------------------------------------------------
@@ -35,7 +31,7 @@ sapply(out, dir.create, showWarnings = TRUE)
 # Collect individual default outputs files to decouple() test.
 
 out_default <- stringr::str_glue(
-    "output-{available_statistics}_dorothea_default.rds"
+    "output-{available_statistics}.rds"
 ) %>%
     map2(.x = out, .y = ., file.path)
 
@@ -43,55 +39,51 @@ decouple_dir <- file.path(output_dir, "decouple")
 dir.create(decouple_dir, showWarnings = TRUE)
 
 # Load data to generated outputs -----------------------------------------------
-
-emat <- file.path(input_dir, "input-expr_matrix.rds") %>%
-    readRDS()
-
-dorothea_genesets <- file.path(input_dir, "input-dorothea_genesets.rds") %>%
-    readRDS()
+mat <- readRDS(file.path(input_dir, 'mat.rds'))
+net <- readRDS(file.path(input_dir, 'net.rds'))
 
 #----- run_udt() ---------------------------------------------------------------
-run_udt(emat, dorothea_genesets, .source='tf') %>%
+run_udt(mat, net, minsize = 0) %>%
     saveRDS(out_default$udt)
 
 #----- run_mdt() ---------------------------------------------------------------
-run_mdt(emat, dorothea_genesets, .source='tf', trees=1000) %>%
+run_mdt(mat, net, minsize=0, trees=1000) %>%
     saveRDS(out_default$mdt)
 
 #----- run_aucell() ------------------------------------------------------------
-run_aucell(emat, dorothea_genesets, .source='tf', nproc=1) %>%
+run_aucell(mat, net, minsize=0, nproc=1, aucMaxRank=3) %>%
     saveRDS(out_default$aucell)
 
 #----- run_wmean() -------------------------------------------------------------
-run_wmean(emat, dorothea_genesets, .source='tf') %>%
+run_wmean(mat, net, minsize=0) %>%
     saveRDS(out_default$wmean)
 
 #----- run_wsum() --------------------------------------------------------------
-run_wsum(emat, dorothea_genesets, .source='tf') %>%
+run_wsum(mat, net, minsize=0) %>%
     saveRDS(out_default$wsum)
 
 #----- run_ulm() ---------------------------------------------------------------
-run_ulm(emat, dorothea_genesets, .source='tf') %>%
+run_ulm(mat, net, minsize=0) %>%
     saveRDS(out_default$ulm)
 
 #----- run_mlm() ---------------------------------------------------------------
-run_mlm(emat, dorothea_genesets, .source='tf') %>%
+run_mlm(mat, net, minsize=0) %>%
     saveRDS(out_default$mlm)
 
 #----- run_viper() -------------------------------------------------------------
-run_viper(emat, dorothea_genesets, .source='tf') %>%
+run_viper(mat, net, minsize=0) %>%
     saveRDS(out_default$viper)
 
 #----- run_gsva() --------------------------------------------------------------
-run_gsva(emat, dorothea_genesets, .source='tf') %>%
+run_gsva(mat, net, minsize=0) %>%
     saveRDS(out_default$gsva)
 
 #----- run_ora() ---------------------------------------------------------------
-run_ora(emat, dorothea_genesets, .source='tf', n_up=300, n_bottom=300) %>%
+run_ora(mat, net, minsize=0, n_up=3, n_bottom=3) %>%
     saveRDS(out_default$ora)
 
 #----- run_fgsea() -------------------------------------------------------------
-run_fgsea(emat, dorothea_genesets, .source='tf') %>%
+run_fgsea(mat, net, minsize=0) %>%
     saveRDS(out_default$fgsea)
 
 # decouple() --------------------------------------------------------------
@@ -101,4 +93,4 @@ run_fgsea(emat, dorothea_genesets, .source='tf') %>%
 
 map_dfr(out_default, readRDS) %>%
     dplyr::arrange(.data$statistic, .data$source, .data$condition) %>%
-    saveRDS(file.path(decouple_dir, "output-decouple_dorothea_default.rds"))
+    saveRDS(file.path(decouple_dir, "output-decouple.rds"))